Move form item base to new file
